requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.2
pandas==2.0.3
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.7.2
pycryptodome==3.17.0
Pillow==9.5.0
tensorboard==2.9.1
